# webshop-backend/internal/handler/return_handler.go
package handler

import (
	"errors"
	"net/http"
	"webshop-backend/internal/models"
	"webshop-backend/internal/repository"

	"github.com/gin-gonic/gin"
)

type ReturnHandler struct {
	repo *repository.ReturnRepository
}

func NewReturnHandler(repo *repository.ReturnRepository) *ReturnHandler {
	return &ReturnHandler{repo: repo}
}

// Create godoc
// @Summary     Submit a return request
// @Tags        returns
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Param       body body     models.CreateReturnRequest true "Return data"
// @Success     201  {object} models.Return
// @Failure     400  {object} map[string]string
// @Failure     404  {object} map[string]string
// @Router      /returns [post]
func (h *ReturnHandler) Create(c *gin.Context) {
	var req models.CreateReturnRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	ret, err := h.repo.Create(userID(c), req)
	if err != nil {
		if errors.Is(err, repository.ErrOrderNotOwned) {
			c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
			return
		}
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, ret)
}

// ListMine godoc
// @Summary     List my return requests
// @Tags        returns
// @Produce     json
// @Security    BearerAuth
// @Success     200 {array}  models.Return
// @Router      /returns [get]
func (h *ReturnHandler) ListMine(c *gin.Context) {
	returns, err := h.repo.ListForUser(userID(c))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	if returns == nil {
		returns = []models.Return{}
	}
	c.JSON(http.StatusOK, returns)
}

// ListAll godoc
// @Summary     List all return requests (admin)
// @Tags        admin
// @Produce     json
// @Security    BearerAuth
// @Success     200 {array}  models.Return
// @Router      /admin/returns [get]
func (h *ReturnHandler) ListAll(c *gin.Context) {
	returns, err := h.repo.ListAll()
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	if returns == nil {
		returns = []models.Return{}
	}
	c.JSON(http.StatusOK, returns)
}

// UpdateStatus godoc
// @Summary     Update return status (admin)
// @Tags        admin
// @Accept      json
// @Security    BearerAuth
// @Param       id   path int                                true "Return ID"
// @Param       body body models.UpdateReturnStatusRequest   true "New status"
// @Success     204
// @Failure     404 {object} map[string]string
// @Router      /admin/returns/{id}/status [put]
func (h *ReturnHandler) UpdateStatus(c *gin.Context) {
	id, err := parseID(c)
	if err != nil {
		return
	}
	var req models.UpdateReturnStatusRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	ok, err := h.repo.UpdateStatus(id, req.Status)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	if !ok {
		c.JSON(http.StatusNotFound, gin.H{"error": "return not found"})
		return
	}
	c.Status(http.StatusNoContent)
}